Rewrote GetChangesFromDisk
GetChangesFromDisk:
Logic error cos i was comparing changes using filePath first (which can change on rename), which made the whole logic confusing. ADDing a new file with the same filePath as a deleted file will result in a MODIFIED. Changed whole system to use LocalId to search through instead of filePath. Also removed parity checking in this section cos that's what SyncEventConflictResolver is for!

SyncEventConflictResolver:
WIP. Did not allow for new files of the same name to be added after deleting a file.